requirements.txt
streamlit
pandas
pandas-ta
python-binance
requests